refactor code

+++ app/src/main/java/com/example/bookapplication/models/catalogmodel/VolumeInfo.kt
@@ -8,7 +8,7 @@ data class VolumeInfo(
     @SerializedName("authors")
     val authors: List<String>,
     @SerializedName("averageRating")
-    val averageRating: Int,
+    val averageRating: Double,
     @SerializedName("canonicalVolumeLink")
     val canonicalVolumeLink: String,
     @SerializedName("categories")
